+++ config/license.php
@@ -1,0 +1,10 @@
+<?php
+
+declare(strict_types=1);
+
+return [
+    'license_serial' => '',
+    'license_api_url' => 'https://adminlicence.obi2.net/api/v1',
+    'license_heartbeat_enabled' => true,
+    'license_heartbeat_interval' => 86400,
+];